Move indexed Persons formatting from TextUi to Formatter
For LO-SRP.
